pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cognisn-spark"
dynamic = ["version"]
description = "Secure Personal AI Research Kit - Multi-provider LLM web interface with MCP tool integration"
readme = "README.md"
license = "MIT with Commons Clause"
requires-python = ">=3.12"
authors = [
    { name = "Matthew Westwood-Hill" },
]
keywords = [
    "llm",
    "ai",
    "chatbot",
    "aws",
    "bedrock",
    "anthropic",
    "claude",
    "ollama",
    "mcp",
    "model-context-protocol",
    "web",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Communications :: Chat",
    "Typing :: Typed",
]
dependencies = [
    "cognisn-konfig>=0.1.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "sse-starlette>=1.6.0",
    "httpx>=0.24.0",
    "aiohttp>=3.8.0",
    "pyyaml>=6.0",
    "tiktoken>=0.5.0",
    "cryptography>=42.0.0",
    # LLM providers
    "anthropic>=0.18.0",
    "boto3>=1.28.0",
    "botocore>=1.31.0",
    "google-genai>=1.0.0",
    "ollama>=0.2.0",
    "openai>=1.0.0",
    # MCP
    "mcp>=0.9.0",
    # Document handling
    "python-docx>=0.8.11",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.21",
    "pdfplumber>=0.10.0",
    "python-magic-bin>=0.4.14; sys_platform == 'win32'",
    "python-magic>=0.4.27; sys_platform != 'win32'",
    # Semantic search and embeddings
    "numpy>=1.21.0",
    "sentence-transformers>=2.2.0",
    # Web tools
    "beautifulsoup4>=4.12.0",
    "html2text>=2024.2.26",
    # Scheduling
    "APScheduler>=3.10.0",
    # System tray
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
    # Markdown
    "markdown>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.8.0",
    "black>=24.0.0",
    "isort>=5.13.0",
]
mysql = ["mysql-connector-python>=8.0.0"]
postgresql = ["psycopg2-binary>=2.9.0"]
mssql = ["pyodbc>=4.0.0"]
all-databases = [
    "mysql-connector-python>=8.0.0",
    "psycopg2-binary>=2.9.0",
    "pyodbc>=4.0.0",
]

[project.scripts]
spark = "spark.launch:main"
spark-daemon = "spark.daemon.tray:run_tray_daemon"

[project.urls]
Homepage = "https://github.com/Cognisn/spark"
Documentation = "https://github.com/Cognisn/spark#readme"
Repository = "https://github.com/Cognisn/spark"
Issues = "https://github.com/Cognisn/spark/issues"
Changelog = "https://github.com/Cognisn/spark/blob/main/CHANGELOG.md"

[tool.hatch.version]
path = "src/spark/_version.txt"
pattern = "(?P<version>.+)"

[tool.hatch.build.targets.wheel]
packages = ["src/spark"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build.targets.sdist]
include = [
    "src/spark/",
    "tests/",
    "README.md",
    "LICENSE",
    "CHANGELOG.md",
    "pyproject.toml",
]

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --tb=short"
